Move useFetch to kolibri-common
