+++ kolibri/plugins/coach/frontend/views/common/resourceSelection/subPages/SelectFromTopicTree.vue
@@ -88,7 +88,7 @@
   import { SelectionTarget } from '../contants';
 
   /**
-   * @typedef {import('../../../../composables/useFetch').FetchObject} FetchObject
+   * @typedef {import('kolibri-common/composables/useFetch.js').FetchObject} FetchObject
    */
 
   export default {
