+++ kolibri/plugins/coach/frontend/views/lessons/LessonSummaryPage/sidePanels/LessonResourceSelection/subPages/SearchFilters.vue
@@ -24,7 +24,7 @@
   import { PageNames } from '../../../../../../constants';
 
   /**
-   * @typedef {import('../../../../../../composables/useFetch').FetchObject} FetchObject
+   * @typedef {import('kolibri-common/composables/useFetch.js').FetchObject} FetchObject
    */
 
   export default {
